Struct conmon_client::unix::UnixStream

pub struct UnixStream {
    inner: Arc<Async<UnixStream>>,
    readable: Option<ReadableOwned<UnixStream>>,
    writable: Option<WritableOwned<UnixStream>>,
}

A Unix connection.

A UnixStream can be created by connecting to an endpoint or by accepting an incoming connection.

UnixStream is a bidirectional stream that implements traits [AsyncRead] and [AsyncWrite].

Cloning a UnixStream creates another handle to the same socket. The socket will be closed when all handles to it are dropped. The reading and writing portions of the connection can also be shut down individually with the shutdown() method.

Examples

use async_net::unix::UnixStream;
use futures_lite::prelude::*;

let mut stream = UnixStream::connect("/tmp/socket").await?;
stream.write_all(b"hello").await?;

let mut buf = vec![0u8; 1024];
let n = stream.read(&mut buf).await?;

Fields
inner: Arc<Async<UnixStream>>readable: Option<ReadableOwned<UnixStream>>writable: Option<WritableOwned<UnixStream>>
Implementations

impl UnixStream

pub async fn connect<P>(path: P) -> Result<UnixStream, Error> where
P: AsRef<Path>,

Creates a Unix connection to given path.

Examples

use async_net::unix::UnixStream;

let stream = UnixStream::connect("/tmp/socket").await?;

pub fn pair() -> Result<(UnixStream, UnixStream), Error>

Creates a pair of connected Unix sockets.

Examples

use async_net::unix::UnixStream;

let (stream1, stream2) = UnixStream::pair()?;

pub fn local_addr(&self) -> Result<SocketAddr, Error>

Returns the local address this socket is connected to.

Examples

use async_net::unix::UnixStream;

let stream = UnixStream::connect("/tmp/socket").await?;
println!("Local address is {:?}", stream.local_addr()?);

pub fn peer_addr(&self) -> Result<SocketAddr, Error>

Returns the remote address this socket is connected to.

Examples

use async_net::unix::UnixStream;

let stream = UnixStream::connect("/tmp/socket").await?;
println!("Connected to {:?}", stream.peer_addr()?);

pub fn shutdown(&self, how: Shutdown) -> Result<(), Error>

Shuts down the read half, write half, or both halves of this connection.

This method will cause all pending and future I/O in the given directions to return immediately with an appropriate value (see the documentation of Shutdown).

use async_net::{Shutdown, unix::UnixStream};

let stream = UnixStream::connect("/tmp/socket").await?;
stream.shutdown(Shutdown::Both)?;